chore: moves eval_framework inside the cognee lib
