+++ cognee/eval_framework/benchmark_adapters/hotpot_qa_adapter.py
@@ -3,13 +3,15 @@ import os
 import json
 import random
 from typing import Optional, Union, Any, LiteralString
-from evals.eval_framework.benchmark_adapters.base_benchmark_adapter import BaseBenchmarkAdapter
+from cognee.eval_framework.benchmark_adapters.base_benchmark_adapter import BaseBenchmarkAdapter
 
 
-class TwoWikiMultihopAdapter(BaseBenchmarkAdapter):
+class HotpotQAAdapter(BaseBenchmarkAdapter):
     dataset_info = {
-        "filename": "2wikimultihop_dev.json",
-        "URL": "https://huggingface.co/datasets/voidful/2WikiMultihopQA/resolve/main/dev.json",
+        "filename": "hotpot_benchmark.json",
+        "url": "http://curtis.ml.cmu.edu/datasets/hotpot/hotpot_dev_distractor_v1.json",
+        # train: "http://curtis.ml.cmu.edu/datasets/hotpot/hotpot_train_v1.1.json" delete file after changing the url
+        # distractor test: "http://curtis.ml.cmu.edu/datasets/hotpot/hotpot_dev_distractor_v1.json" delete file after changing the url
     }
 
     def load_corpus(
@@ -21,7 +23,7 @@ class TwoWikiMultihopAdapter(BaseBenchmarkAdapter):
             with open(filename, "r", encoding="utf-8") as f:
                 corpus_json = json.load(f)
         else:
-            response = requests.get(self.dataset_info["URL"])
+            response = requests.get(self.dataset_info["url"])
             response.raise_for_status()
             corpus_json = response.json()
 
@@ -34,15 +36,15 @@ class TwoWikiMultihopAdapter(BaseBenchmarkAdapter):
 
         corpus_list = []
         question_answer_pairs = []
-        for dict in corpus_json:
-            for title, sentences in dict["context"]:
+        for item in corpus_json:
+            for title, sentences in item["context"]:
                 corpus_list.append(" ".join(sentences))
 
             question_answer_pairs.append(
                 {
-                    "question": dict["question"],
-                    "answer": dict["answer"].lower(),
-                    "type": dict["type"],
+                    "question": item["question"],
+                    "answer": item["answer"].lower(),
+                    "level": item["level"],
                 }
             )
